% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_evaluate.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy}
\usage{
accuracy(pred, true, ...)
}
\arguments{
\item{pred}{vector of predicted labels derived from some model, such as \link{textmodel_mlp},
which is being subjected to evaluation}

\item{true}{a vector of known labels that are used to evaluate model performance}

\item{...}{additional arguments}
}
\description{
\code{accuracy()} Implements a function that calculates the proportion of positive and negative values that have been correctly identified by a machine learning model and a vector of true values.
It is calculated by calculating the quotient of the total correctly identified labels and the total number of labels. Its values range from 0 to 1, where 0 would indicate that no predicted labels were correct
and 1 would indicate that all predicted labels were correct.
}
\seealso{
\code{\link[=textmodel_evaluate]{textmodel_evaluate()}}
}
