% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_evaluate.R
\name{balanced_accuracy}
\alias{balanced_accuracy}
\title{Balanced Accuracy}
\usage{
balanced_accuracy(pred, true, by_class = FALSE, ...)
}
\arguments{
\item{pred}{vector of predicted labels derived from some model, such as \link{textmodel_mlp},
which is being subjected to evaluation}

\item{true}{a vector of known labels that are used to evaluate model performance}

\item{by_class}{instead of returning the average accuracy across all labels, this function will return the balanced accuracy score for every class of the true vector}

\item{...}{additional arguments}
}
\description{
\code{accuracy()} Implements a function that calculates the average proportion of positive and negative values that have been correctly identified within each class by a machine learning model and a vector of true values.
It is calculated by first calculating the proportion of correct predictions within each label and then taking the mean of the resulting vector. Its values range from 0 to 1, where 0 would indicate that the average proportion
of correctly classified labels is 0 and 1 would indicate that the average proportion of correctly classified labels is 1.
}
\seealso{
\code{\link[=textmodel_evaluate]{textmodel_evaluate()}}
}
