% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_corpus_LMRD}
\alias{data_corpus_LMRD}
\title{Large Movie Review Dataset from Maas et. al. (2011)}
\format{
The corpus docvars consist of:
\describe{
\item{docnumber}{serial (within set and polarity) document number}
\item{rating}{user-assigned movie rating on a 1-10 point integer scale}
\item{set}{used for test v. training set}
\item{polarity}{either \code{neg} or \code{pos} to indicate whether the
movie review was negative or positive.  See Maas et al (2011) for the
cut-off values that governed this assignment.}
}
}
\source{
\url{http://ai.stanford.edu/~amaas/data/sentiment/}
}
\usage{
data_corpus_LMRD
}
\description{
A corpus object containing a dataset for sentiment classification containing
25,000 highly polar movie reviews for training, and 25,000 for testing, from
Maas et. al. (2011).
}
\references{
Andrew L. Maas, Raymond E. Daly, Peter T. Pham, Dan Huang, Andrew
Y. Ng, and Christopher Potts. (2011).
"\href{http://ai.stanford.edu/~amaas/papers/wvSent_acl2011.pdf}{Learning Word Vectors for Sentiment Analysis}". The 49th Annual Meeting of the
Association for Computational Linguistics (ACL 2011).
}
\keyword{data}
