% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens2sequences.R
\name{docnames.tokens2sequences}
\alias{docnames.tokens2sequences}
\title{Returns docnames from an tokens2sequences or textmodel_cnnlstmemb object}
\usage{
\method{docnames}{tokens2sequences}(x, ...)
}
\arguments{
\item{x}{output from \code{\link[=tokens2sequences]{tokens2sequences()}}}

\item{...}{additional arguments not used}
}
\description{
Returns docnames from an tokens2sequences or textmodel_cnnlstmemb object
}
\keyword{textmodel}
