% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performance}
\alias{performance}
\alias{precision}
\alias{recall}
\alias{f1_score}
\alias{accuracy}
\alias{balanced_accuracy}
\title{Performance statistics for prediction}
\usage{
performance(data, truth, by_class = TRUE, ...)

precision(data, truth, by_class = TRUE, ...)

recall(data, truth, by_class = TRUE, ...)

f1_score(data, truth, by_class = TRUE, ...)

accuracy(data, truth, ...)

balanced_accuracy(data, ...)
}
\arguments{
\item{data}{a table of predicted by truth, or vector of predicted labels}

\item{truth}{vector of "true" labels, or if a table, \code{2} to indicate that the
"true" values are in columns, or \code{1} if in rows.}

\item{by_class}{logical; if \code{TRUE}, estimate performance score separately for
each class, otherwise average across classes}

\item{...}{not used}
}
\value{
named list consisting of the selected measure(s), where each element
is a scalar if \code{by_class = FALSE}, or a vector named by class if \code{by_class = TRUE}.
}
\description{
Functions for computing performance statistics used for model
evaluation.

\code{performance()} computes all of the following, which are also
available via specific functions:

Given a 2 x 2 table with notation

\tabular{rcc}{ \tab Truth \tab \cr Predicted \tab Positive \tab
Negative \cr Positive \tab \emph{A} \tab \emph{B} \cr Negative \tab \emph{C} \tab \emph{D} \cr }

The metrics computed here are:
\itemize{
\item{precision: }{\eqn{A / (A + B)}}
\item{recall: }{\eqn{A / (A + C)}}
\item{\emph{F1}: }{\eqn{2 / (recall^{-1} + precision^{-1})}}
\item{accuracy: }{\eqn{(A + D) / (A + B + C + D)}, or correctly predicted / all}
\item{balanced_accuracy: }{mean(recall) for all categories}
}
}
\examples{
## Data in Table 2 of Powers (2007)

lvs <- c("Relevant", "Irrelevant")
tbl_2_1_pred <- factor(rep(lvs, times = c(42, 58)), levels = lvs)
tbl_2_1_truth <- factor(c(rep(lvs, times = c(30, 12)),
                          rep(lvs, times = c(30, 28))),               
                        levels = lvs)
                        
performance(tbl_2_1_pred, tbl_2_1_truth)
performance(tbl_2_1_pred, tbl_2_1_truth, by_class = FALSE)
performance(table(tbl_2_1_pred, tbl_2_1_truth), by_class = TRUE)

precision(tbl_2_1_pred, tbl_2_1_truth)

recall(tbl_2_1_pred, tbl_2_1_truth)

f1_score(tbl_2_1_pred, tbl_2_1_truth)

accuracy(tbl_2_1_pred, tbl_2_1_truth)

balanced_accuracy(tbl_2_1_pred, tbl_2_1_truth)

}
\references{
Powers, D. (2007). "Evaluation: From Precision, Recall and F Factor to ROC,
Informedness, Markedness and Correlation." \emph{Technical Report SIE-07-001},
Flinders University.
}
