% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_mlp.R
\name{predict.textmodel_mlp}
\alias{predict.textmodel_mlp}
\title{Prediction from a fitted textmodel_mlp object}
\usage{
\method{predict}{textmodel_mlp}(
  object,
  newdata = NULL,
  type = c("class", "probability"),
  force = TRUE,
  ...
)
}
\arguments{
\item{object}{a fitted \link{textmodel_mlp} model}

\item{newdata}{dfm on which prediction should be made}

\item{type}{the type of predicted values to be returned; see Value}

\item{force}{make \code{newdata}'s feature set conformant to the model terms}

\item{...}{not used}
}
\value{
\code{predict.textmodel_mlp} returns either a vector of class
predictions for each row of \code{newdata} (when \code{type = "class"}), or
a document-by-class matrix of class probabilities (when \code{type = "probability"}).
}
\description{
\code{predict.textmodel_mlp()} implements class predictions from a fitted
multilayer perceptron network model.
}
\seealso{
\code{\link[=textmodel_mlp]{textmodel_mlp()}}
}
\keyword{internal}
\keyword{textmodel}
