% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_seq.R
\name{predict.textmodel_seq}
\alias{predict.textmodel_seq}
\title{Prediction from a fitted textmodel_seq object}
\usage{
\method{predict}{textmodel_seq}(object, newdata = NULL,
  type = c("class", "probability"), force = TRUE, ...)
}
\arguments{
\item{object}{a fitted linear SEQ textmodel}

\item{newdata}{dfm on which prediction should be made}

\item{type}{the type of predicted values to be returned; see Value}

\item{force}{make newdata's feature set conformant to the model terms}

\item{...}{not used}
}
\value{
\code{predict.textmodel_seq} returns either a vector of class
  predictions for each row of \code{newdata} (when \code{type = "class"}), or
  a document-by-class matrix of class probabilities (when \code{type =
  "probability"}).
}
\description{
\code{predict.textmodel_seq()} implements class predictions from a fitted
SEQ model.
}
\seealso{
\code{\link{textmodel_seq}}
}
\keyword{internal}
\keyword{textmodel}
