% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_slstm.R
\name{predict.textmodel_slstm}
\alias{predict.textmodel_slstm}
\title{Prediction from a fitted textmodel_slstm object}
\usage{
\method{predict}{textmodel_slstm}(object, newdata = NULL,
  type = c("class", "probability"), force = TRUE, ...)
}
\arguments{
\item{object}{a fitted \link{textmodel_slstm} model}

\item{newdata}{dfm on which prediction should be made}

\item{type}{the type of predicted values to be returned; see Value}

\item{force}{make \code{newdata}'s feature set conformant to the model terms}

\item{...}{not used}
}
\value{
\code{predict.textmodel_slstm} returns either a vector of class
  predictions for each row of \code{newdata} (when \code{type = "class"}), or
  a document-by-class matrix of class probabilities (when \code{type =
  "probability"}).
}
\description{
\code{predict.textmodel_slstm()} implements class predictions from a fitted
sequential neural network model.
}
\seealso{
\code{\link{textmodel_slstm}}
}
\keyword{internal}
\keyword{textmodel}
