% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda.classifiers-package.R
\docType{package}
\name{quanteda.classifiers-package}
\alias{quanteda.classifiers}
\alias{quanteda.classifiers-package}
\title{quanteda.classifiers}
\description{
Extensions to \pkg{quanteda} that provide supervised machine learning models
for document-feature matrices.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/quanteda/quanteda.svm}
  \item Report bugs at \url{https://github.com/quanteda/quanteda.svm/issues}
}

}
\author{
\strong{Maintainer}: Kenneth Benoit \email{kbenoit@lse.ac.uk}

Authors:
\itemize{
  \item Patrick Chester \email{pjc468@nyu.edu}
  \item Müller Stefan \email{mueller@ipz.uzh.ch}
}

}
\keyword{internal}
