% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_evaluate.R
\name{recall}
\alias{recall}
\title{Recall}
\usage{
recall(pred, true, by_class = FALSE, ...)
}
\arguments{
\item{pred}{vector of predicted labels derived from some model, such as \link{textmodel_mlp},
which is being subjected to evaluation}

\item{true}{a vector of known labels that are used to evaluate model performance}

\item{by_class}{estimates a separate recall score for every class of the true vector}

\item{...}{additional arguments}
}
\description{
\code{recall()} Implements a function that calculates the recall, also known as the sensitivity, of labels predicted by a machine learning model and a vector of true values.
It is calculated by calculating the ratio of true predicted positives to total true positives. Its values range from 0 to 1, where 0 would indicate that no true positives
were predicted by the model while 1 would indicate that all true positives in the data were identified.
}
\seealso{
\code{\link[=textmodel_evaluate]{textmodel_evaluate()}}
}
