% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_mlp.R
\name{summary.textmodel_mlp}
\alias{summary.textmodel_mlp}
\title{summary method for textmodel_mlp objects}
\usage{
\method{summary}{textmodel_mlp}(object, ...)
}
\arguments{
\item{object}{output from \code{\link[=textmodel_mlp]{textmodel_mlp()}}}

\item{...}{additional arguments not used}
}
\description{
summary method for textmodel_mlp objects
}
\keyword{internal}
\keyword{textmodel}
