% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_nnseq.R
\name{summary.textmodel_nnseq}
\alias{summary.textmodel_nnseq}
\title{summary method for textmodel_svm objects}
\usage{
\method{summary}{textmodel_svm}(object, ...)
}
\arguments{
\item{object}{output from \code{\link{textmodel_svm}}}

\item{...}{additional arguments not used}

\item{n}{how many coefficients to print before truncating}
}
\description{
summary method for textmodel_svm objects
}
\keyword{internal}
\keyword{textmodel}
