% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens2sequences.R
\name{t2s_conform}
\alias{t2s_conform}
\title{Converts the feature names of one text2sequences object to that of another}
\usage{
\method{t2s_conform}{tokens2sequences}(x, y)
}
\arguments{
\item{x}{tokens2sequence object that will be forced to conform}

\item{y}{tokens2sequence object whose feature names will be used to change token labels for \code{x}}
}
\description{
Converts the feature names of one text2sequences object to that of another
}
\examples{
\dontrun {
corpcoded <- corpus_subset(data_corpus_manifestosentsUK, !is.na(crowd_immigration_label))
corpuncoded <- data_corpus_manifestosentsUK \%>\%
    corpus_subset(is.na(crowd_immigration_label) & year > 1980) \%>\%
    corpus_sample(size = ndoc(corpcoded))

tokx <- tokens(texts(corpuncoded))
toky <- tokens(texts(corpcoded))

seqx <- tokens2sequences(tokx, maxsenlen = 50, keepn = 5000)
seqy <- tokens2sequences(toky, maxsenlen = 50, keepn = 5000)

seqxy <- t2s_conform(seqx, seqy)

print(seqxy)
}
}
\seealso{
\code{\link{tokens2sequences}}
}
