% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_cnnlstmemb.R
\name{test_embeddings}
\alias{test_embeddings}
\title{Test whether embeddings are of the correct format: n x k + 1,
where the rows are n tokens and columns are the k dimensions. The first column should be the token labels.}
\usage{
test_embeddings(matrix)
}
\arguments{
\item{matrix}{Embedding matrix}
}
\description{
Test whether embeddings are of the correct format: n x k + 1,
where the rows are n tokens and columns are the k dimensions. The first column should be the token labels.
}
