% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_seq.R
\name{textmodel_seq}
\alias{textmodel_seq}
\title{Neural Network with Two Layers}
\usage{
textmodel_seq(x, y, Seed = 17, Epochs = 3, Units = 512, Batch = 32,
  Dropout = 0.2, Valsplit = 0.1, Metric = "categorical_accuracy",
  Loss = "categorical_crossentropy", Optimizer = "adam",
  Verbose = TRUE)
}
\arguments{
\item{x}{The data that will be used as training and test data.}

\item{y}{The outcome that will be used as outcomes to be predicted by the NN model.}

\item{Seed}{The seed used in the model. Defaults to 17}

\item{Epochs}{The number of epochs used in the NN model.}

\item{Units}{The number of network nodes used in the first layer of the sequential model}

\item{Batch}{The number of batches estimated}

\item{Dropout}{A floating variable bound between 0 and 1. It determines the rate at which units are dropped for the linear tranformation of the inputs.}

\item{Metric}{Metric used to train algorithm}

\item{Loss}{Metric used to train algorithm}

\item{Optimizer}{Optimizer used to fit model to training data}

\item{Weighting}{The type of feature weighting used in the document feature matrix. I.e., count and tfidf.}

\item{ValSplit}{The validation split of the data used in the training of the LSTM model}
}
\description{
This function is a wrapper for a two layered neural network written using the Keras Package. It takes a
}
\keyword{networks}
\keyword{neural}
