% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens2sequences.R
\name{tokens2sequences}
\alias{tokens2sequences}
\title{convert tokens to sequences for keras word2vec input}
\usage{
tokens2sequences(x)
}
\arguments{
\item{x}{}
}
\value{
explain return format
}
\description{
More description here, with links to appropriate \pkg{keras} functions.
}
\examples{
# add one
}
