% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens2sequences.R
\name{tokens2sequences_convert}
\alias{tokens2sequences_convert}
\title{Convert a matrix and label dictionary to a \code{\link[=tokens2sequences]{tokens2sequences()}} object}
\usage{
tokens2sequences_convert(mat, dict)
}
\arguments{
\item{mat}{A matrix containing numeric representations of tokens of dimensions n x k, where n is the document level and k represents the maximum
sequence length}

\item{dict}{A data frame with one numeric column representing}
}
\description{
Convert a matrix and label dictionary to a \code{\link[=tokens2sequences]{tokens2sequences()}} object
}
