% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossval}
\alias{crossval}
\title{Cross-validate a fitted textmodel}
\usage{
crossval(x, k = 5, by_class = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{a fitted textmodel}

\item{k}{number of folds}

\item{by_class}{logical; if \code{TRUE}, estimate performance score separately for
each class, otherwise average across classes}

\item{verbose}{logical; if \code{TRUE}, output results to the console}
}
\description{
Cross-validate a fitted textmodel using \emph{k}-fold cross-validation.
}
\examples{
library("quanteda")
library("quanteda.textmodels")
dfmat <- dfm(data_corpus_moviereviews)
textmodel_nb(dfmat, y = data_corpus_moviereviews$sentiment) \%>\%
    crossval(k = 5, by_class = TRUE)
textmodel_nb(dfmat, y = data_corpus_moviereviews$sentiment) \%>\%
    crossval(k = 5, by_class = FALSE)
textmodel_svmlin(dfmat, y = data_corpus_moviereviews$sentiment) \%>\%
    crossval(k = 5, by_class = TRUE)
}
