% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokens2sequences.R
\name{remove_features}
\alias{remove_features}
\title{Function that applies feature reduction to data set}
\usage{
remove_features(x, data, maxsenlen, keepn, keep_beginning)
}
\arguments{
\item{x}{List of token IDs}

\item{data}{Data frame that maps features to their IDs}

\item{maxsenlen}{the maximum sentence length kept in output matrix}

\item{keepn}{the maximum number of features to keep}

\item{keep_beginning}{Boolean indicating whether tokens kept are
those at the beginning of the text sequence or the end}
}
\description{
Function that applies feature reduction to data set
}
